New hook for services
New hook for services, code refactoring and some UI changes
