--- src/services.json
@@ -1,30 +1,0 @@
-[
-  {
-    "name": "Aloitus 1€",
-    "pricePerMin": 0.25
-  },
-  {
-    "name": "Aloitus 1€",
-    "pricePerMin": 0.24
-  },
-  {
-    "name": "Aloitus 1€",
-    "pricePerMin": 0.23
-  },
-  {
-    "name": "Aloitus 1€",
-    "pricePerMin": 0.22
-  },
-  {
-    "name": "Aloitus 1€",
-    "pricePerMin": 0.21
-  },
-  {
-    "name": "Aloitus 1€",
-    "pricePerMin": 0.2
-  },
-  {
-    "name": "Aloitus 1€",
-    "pricePerMin": 0.19
-  }
-]
